% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes10-1}
\alias{barcodes10-1}
\title{barcodes10-1}
\format{An object of class \code{character} of length 1903.}
\usage{
barcodes10-1

barcodes10-1

barcodes10-1
}
\description{
barcodes10-1

barcodes10-1

barcodes10-1
}
\keyword{datasets}
