% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes12-2}
\alias{barcodes12-2}
\title{barcodes12-2}
\format{An object of class \code{character} of length 179.}
\usage{
barcodes12-2

barcodes12-2

barcodes12-2
}
\description{
barcodes12-2

barcodes12-2

barcodes12-2
}
\keyword{datasets}
