% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes13-1}
\alias{barcodes13-1}
\title{barcodes13-1}
\format{An object of class \code{character} of length 56736.}
\usage{
barcodes13-1

barcodes13-1

barcodes13-1
}
\description{
barcodes13-1

barcodes13-1

barcodes13-1
}
\keyword{datasets}
