% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes14-1}
\alias{barcodes14-1}
\title{barcodes14-1}
\format{An object of class \code{character} of length 157197.}
\usage{
barcodes14-1

barcodes14-1

barcodes14-1
}
\description{
barcodes14-1

barcodes14-1

barcodes14-1
}
\keyword{datasets}
