% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes15-1}
\alias{barcodes15-1}
\title{barcodes15-1}
\format{An object of class \code{character} of length 518509.}
\usage{
barcodes15-1

barcodes15-1

barcodes15-1
}
\description{
barcodes15-1

barcodes15-1

barcodes15-1
}
\keyword{datasets}
