% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes16-2}
\alias{barcodes16-2}
\title{barcodes16-2}
\format{An object of class \code{character} of length 8777.}
\usage{
barcodes16-2

barcodes16-2

barcodes16-2
}
\description{
barcodes16-2

barcodes16-2

barcodes16-2
}
\keyword{datasets}
