% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes4-1}
\alias{barcodes4-1}
\title{barcodes4-1}
\format{An object of class \code{character} of length 3.}
\usage{
barcodes4-1

barcodes4-1

barcodes4-1
}
\description{
barcodes4-1

barcodes4-1

barcodes4-1
}
\keyword{datasets}
