% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes5-2}
\alias{barcodes5-2}
\title{barcodes5-2}
\format{An object of class \code{character} of length 2.}
\usage{
barcodes5-2

barcodes5-2

barcodes5-2
}
\description{
barcodes5-2

barcodes5-2

barcodes5-2
}
\keyword{datasets}
