% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes6-1}
\alias{barcodes6-1}
\title{barcodes6-1}
\format{An object of class \code{character} of length 27.}
\usage{
barcodes6-1

barcodes6-1

barcodes6-1
}
\description{
barcodes6-1

barcodes6-1

barcodes6-1
}
\keyword{datasets}
