% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes7-1}
\alias{barcodes7-1}
\title{barcodes7-1}
\format{An object of class \code{character} of length 67.}
\usage{
barcodes7-1

barcodes7-1

barcodes7-1
}
\description{
barcodes7-1

barcodes7-1

barcodes7-1
}
\keyword{datasets}
