% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes8-2}
\alias{barcodes8-2}
\title{barcodes8-2}
\format{An object of class \code{character} of length 7.}
\usage{
barcodes8-2

barcodes8-2

barcodes8-2
}
\description{
barcodes8-2

barcodes8-2

barcodes8-2
}
\keyword{datasets}
