% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes9-1}
\alias{barcodes9-1}
\title{barcodes9-1}
\format{An object of class \code{character} of length 554.}
\usage{
barcodes9-1

barcodes9-1

barcodes9-1
}
\description{
barcodes9-1

barcodes9-1

barcodes9-1
}
\keyword{datasets}
