% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVCFfast.R
\name{processSnp}
\alias{processSnp}
\title{process an individual SNP}
\usage{
processSnp(snp, nper, seqwidth, fwprimer, revprimer)
}
\arguments{
\item{snp}{a data_frame containing the VCF information for one SNP as well as
a barcode pool to sample from.}

\item{nper}{The number of barcoded sequences to be generated per allele per
SNP}

\item{seqwidth}{The amount of sequence context flanking both sides of the SNP
in the generated sequences}

\item{fwprimer}{a string containing the forward PCR primer to be used}

\item{revprimer}{a string containing the reverse PCR primer to be used}
}
\value{
a data_frame of labeled sequences or a data_frame listing the SNP and
  why it failed
}
\description{
Take one SNP, get its genomic context, concatenate the parts, and check it
doesn't contain aberrant digestion sites. If it does, resample the barcodes
and try a few more times. If it still does, return a failure stating why.
}
