% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVCFfast.R
\name{randomly_fix}
\alias{randomly_fix}
\title{Randomly correct aberrant digestion sites}
\usage{
randomly_fix(snp, res_df, dig_patterns, dig_site_locations)
}
\arguments{
\item{snp}{a data_frame of one SNP
#' @param nper The number of barcoded sequences to be generated per allele per
  SNP}

\item{upstreamContextRange}{the amount of sequence context to acquire upstream of the SNP}

\item{downstreamContextRange}{the amount of sequence context to acquire downstream of the SNP}

\item{fwprimer}{a string containing the forward PCR primer to be used}

\item{revprimer}{a string containing the reverse PCR primer to be used}
}
\description{
For a SNP with aberrant digestion sites in the context, randomly change bases
in the site across barcodes
}
