% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVCFfast.R
\name{spread_and_fix_indels}
\alias{spread_and_fix_indels}
\title{Prepare input VCF}
\usage{
spread_and_fix_indels(vcf_path)
}
\arguments{
\item{vcf_path}{path to a vcf to be fixed}
}
\value{
a data_frame of the fixed VCF
}
\description{
Prepare an input VCF for sequence generation
}
\details{
This function takes an input vcf, reads it in, spreads any SNPs with
  alternate alleles across multiple rows, and fixes any indels that are
  improperly formatted. This means if REF and ALT are listed as "A" and "ATC"
  they will be replaced with "-" and "TC". If they're "A" and "T,C" this will
  be spread into two entries of "A" and "T" and "A" and "C".

  The output is written to the same directory as the input named
  "*_fixed.vcf". This is ready to be fed into processVCF()
}
