% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twelvemers}
\alias{twelvemers}
\title{1140292 inert twelvemers}
\format{
An object of class \code{character} of length 1140292.
}
\usage{
twelvemers
}
\description{
These were generated in the designMPRA project, see there for details
https://github.com/andrewGhazi/designMPRA
#source \url{https://github.com/andrewGhazi/designMPRA}
}
\keyword{datasets}
