% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVCFfast.R
\name{processSnp}
\alias{processSnp}
\title{process an individual SNP}
\usage{
processSnp(
  snp,
  nper,
  upstreamContextRange,
  downstreamContextRange,
  fwprimer,
  revprimer,
  enzyme1,
  enzyme2,
  enzyme3,
  alter_aberrant = FALSE,
  extra_elements = FALSE,
  max_construct_size = NULL,
  flip_RV = TRUE
)
}
\arguments{
\item{snp}{a data_frame containing the VCF information for one SNP as well as
a barcode pool to sample from.}

\item{nper}{The number of barcoded sequences to be generated per allele per
SNP}

\item{upstreamContextRange}{the amount of sequence context to acquire upstream of the SNP}

\item{downstreamContextRange}{the amount of sequence context to acquire downstream of the SNP}

\item{fwprimer}{a string containing the forward PCR primer to be used}

\item{revprimer}{a string containing the reverse PCR primer to be used}

\item{enzyme1}{the first restriction enzyme's recognition pattern}

\item{enzyme2}{the first restriction enzyme's recognition pattern}

\item{enzyme3}{the first restriction enzyme's recognition pattern}
}
\value{
a data_frame of labeled sequences with appropriate information on the changes made
}
\description{
Take one SNP, get its genomic context, concatenate the parts, and check it
doesn't contain aberrant digestion sites. If it does, resample the barcodes
and try a few more times. If it still does, return a failure stating why.
}
