% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVCFfast.R
\name{processVCF}
\alias{processVCF}
\title{Process VCF into MPRA sequences}
\usage{
processVCF(vcf, nper, seqwidth, fwprimer, revprimer, outPath = NULL)
}
\arguments{
\item{vcf}{the path to the input VCF}

\item{nper}{The number of barcoded sequences to be generated per allele per
SNP}

\item{seqwidth}{The amount of sequence context flanking both sides of the SNP
in the generated sequences}

\item{fwprimer}{a string containing the forward PCR primer to be used}

\item{revprimer}{a string containing the reverse PCR primer to be used}

\item{outPath}{an optional path stating where to write a .tsv of the results}
}
\value{
A list of two data_frames. The first, named 'result', is a data_frame
  containing the labeled MPRA sequences. The second, named 'failed', is a
  data_frame listing the SNPs that are not able to have MPRA sequences
  generated and the reason why.
}
\description{
\code{processVCF} takes a VCF and returns a data_frame
containing the labeled MPRA sequences barcoded with inert twelvemers
}
